create a parent test class for IT
