+++ src/test/java/com/jobfinder/finder/integrationTest/controller/FinderIntegrationTestInitiator.java
@@ -1,0 +1,46 @@
+package com.jobfinder.finder.integrationTest.controller;
+
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.jobfinder.finder.controller.UserController;
+import com.jobfinder.finder.integrationTest.configuration.MockUserDetailsManagerConfig;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
+
+@Import(MockUserDetailsManagerConfig.class)
+@EnableAutoConfiguration(exclude = {
+    org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration.class,
+    org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration.class,
+    RedisAutoConfiguration.class
+})
+@ActiveProfiles("test")
+public abstract class FinderIntegrationTestInitiator {
+  //this class is used to initiate integration tests for the Finder application
+  @Autowired
+  WebApplicationContext context;
+
+  @Autowired
+  MockMvc mockMvc;
+
+  @Autowired
+  ObjectMapper objectMapper;
+
+  @BeforeEach
+  void setUp() {
+    mockMvc = MockMvcBuilders
+        .webAppContextSetup(context)
+        .apply(springSecurity())
+        .build();
+  }
+
+}
